requirements.txt
krakenex == 0.0.5
python_bittrex == 0.3.0
requests == 2.18.4
